Input parameter finalTime changed to maxLength
Integration of strainlines, shearlines and no stretch lines can stop
using ODE event detection. If this mechanism does not stop
integration, integration is instead stopped with a maximum time. This
time is set heuristically. Since the vector fields of the strainlines,
shearlines and no stretch lines are normalized, a maximum length for
these structures is specified. This should be more intuitive than
specifying a maximum, final integration time.